requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-redis==4.12.1
djangorestframework==3.11.1
idna==2.10
kombu==4.6.11
loadtest==0.0.2
pytz==2020.1
redis==3.5.3
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
